Detect non-legacy tokens using encryption key lenght too.
To interoperate with other implementations that won't put the '__v'
magic key in the headers, it is necessary to enter the non-legacy
code branch using some other mechanism than the presence of '__v'
(which I understand to be an invention of this particular
implementation).

I won't claim to have extensive knowledge or experience with JOSE, but
this trick at least lets me interoperate with a dot-NET client.